refactor: reuse review report types
