+++ frontend/src/pages/user/reviewer/ReviewerPage.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 import { Link } from "react-router-dom";
 import { useToast } from "../../../context/ToastProvider";
 import { getReviewReports } from "../../../api";
-import type { ReviewReport } from "../../../types/reviews.types";
+import type { ReviewReport } from "../../../types/review_reports";
 
 export default function ReviewerPage() {
   const { show } = useToast();
